Description
# blog-bare-bones

1. Copy the folder AppDemo out of the directory and place it into your www folder.
2. check config file for db password *****,
3. Change the password into your relevant db password.
4. Create database `thato_blog`
5. Run the db.sql script to populate your database.
6. Open the project in browser.
7. Login using details
-Username = demo
-Password = demo

*Note: 

I did this as a side project at home.
It's intended to showcase what I do outside of office hours, as my work code is confidential.

a. Admin side with the above details is working.
b. Public user side is a little faulty, will make time to fix it up.

I would have fixed up the bugs and completed it earlier but I then added a more user friendly UI and began converting it to Codeigniter framework.
Will upload that as soon as I am done with bug fixing.